requirements.txt
numpy
opencv-python
pillow
pascal_voc_writer
cvzone
pyyaml
pandas
mediapipe==0.10.18
torch==2.5.0
pyserial
